Introduced options to limit transaction data download to given timeframe

+++ eutl_scraper/pipelines.py
@@ -5,18 +5,20 @@ from eutl_scraper.items import (AccountItem, ContactItem,
                                 SurrenderingDetailsItem, TransactionItem,
                                 TransactionBlockItem, AccountIDMapItem)
 from eutl_scraper.own_settings import DIR_PARSED
+import os.path
+
 
 
 class EutlScraperPipeline:
     itemsToProcess = [
-        {"item": AccountItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'accounts.csv'},
-        {"item": ContactItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'contacts.csv'},
-        {"item": InstallationItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'installations.csv'},
-        {"item": ComplianceItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'compliance.csv'},
-        {"item": SurrenderingDetailsItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'surrendering.csv'},
-        {"item": TransactionItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'transactions.csv'},
-        {"item": TransactionBlockItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'transactionBlocks.csv'},
-        {"item": AccountIDMapItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'accountIdMap.csv'},
+        {"item": AccountItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'accounts.csv', "appendExisting": False},
+        {"item": ContactItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'contacts.csv', "appendExisting": False},
+        {"item": InstallationItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'installations.csv', "appendExisting": False},
+        {"item": ComplianceItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'compliance.csv', "appendExisting": False},
+        {"item": SurrenderingDetailsItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'surrendering.csv', "appendExisting": False},
+        {"item": TransactionItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'transactions.csv', "appendExisting": True},
+        {"item": TransactionBlockItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'transactionBlocks.csv', "appendExisting": True},
+        {"item": AccountIDMapItem, "rows": [], "header": [], "output_file_name": DIR_PARSED + 'accountIdMap.csv', "appendExisting": True},
     ]
 
     def process_item(self, item, spider):
@@ -30,7 +32,15 @@ class EutlScraperPipeline:
         for it in self.itemsToProcess:
             if len(it["rows"]) == 0:
                 continue
-            with open(it["output_file_name"], "w", newline="", encoding="utf-8") as output_file:
-                dict_writer = csv.DictWriter(output_file, it["header"])
-                dict_writer.writeheader()
-                dict_writer.writerows(it["rows"])
+            
+            # check whether we allow to append to existing files
+            if os.path.isfile(it["output_file_name"]) and it["appendExisting"]:
+                with open(it["output_file_name"], "a", newline="", encoding="utf-8") as output_file:
+                    dict_writer = csv.DictWriter(output_file, it["header"])
+                    #dict_writer.writeheader()
+                    dict_writer.writerows(it["rows"])
+            else:
+                with open(it["output_file_name"], "w", newline="", encoding="utf-8") as output_file:
+                    dict_writer = csv.DictWriter(output_file, it["header"])
+                    dict_writer.writeheader()
+                    dict_writer.writerows(it["rows"])
